Responsive menu works
